# docs/api/generated/openapi.yaml
openapi: 3.0.0
info:
  title: Property Management API
  version: 0.0.0
...
paths:
...
  /api/v1/documents/folders:
    get:
      operationId: DocumentsApi_listFolders
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: X-Tenant-ID
          in: header
          required: false
          schema:
            $ref: '#/components/schemas/Shared.uuid'
        - name: parentId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Shared.uuid'
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Documents.Folder'
      tags:
        - Documents
    post:
      operationId: DocumentsApi_createFolder
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: X-Tenant-ID
          in: header
          required: false
          schema:
            $ref: '#/components/schemas/Shared.uuid'
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Documents.Folder'
      tags:
        - Documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                parentId:
                  $ref: '#/components/schemas/Shared.uuid'
                visibility:
                  $ref: '#/components/schemas/Documents.DocumentVisibility'
              required:
                - name
                - visibility
...
components:
...
  schemas:
...
    Documents.DocumentVisibility:
      type: string
      enum:
        - public
        - building_only
        - owners_only
        - managers_only
        - private
    Documents.Folder:
      type: object
      required:
        - id
        - organizationId
        - name
        - visibility
        - documentCount
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/Shared.uuid'
        organizationId:
          $ref: '#/components/schemas/Shared.uuid'
        parentId:
          $ref: '#/components/schemas/Shared.uuid'
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/Documents.DocumentVisibility'
        documentCount:
          type: integer
          format: int32
        createdAt:
          allOf:
            - $ref: '#/components/schemas/Shared.dateTime'
          description: Creation timestamp
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/Shared.dateTime'
          description: Last update timestamp
        createdBy:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: ID of user who created this entity
        updatedBy:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: ID of user who last updated this entity
      description: Document folder
...
    Shared.dateTime:
      type: string
      description: ISO 8601 date-time string
...
    Shared.uuid:
      type: string
      description: Unique identifier (UUID v4)
...
servers:
  - url: http://localhost:8080
    description: Development
    variables: {}
  - url: https://api.property-management.example.com
    description: Production
    variables: {}